SQLens/requirements.txt
streamlite
google-generaiveai
python -dotenv